* develop: (90 commits)
  Fix multiple GPUs with GPU preprocess action
  twig scripts to genrate quantized models, now works at develop
  Update convert script to ensure directory exists
  Fix different number of bboxes may get batched in tf.map_fn
  Confidence as transparency
  Better bbox drawing with transparency
  Uncomment bbox filtering
  Add conversion rules for YOLO
  Finally correct detection with tiny_yolo
  Fix distort_bbox regression
  make stochastic a modular parameter in all quantization methods, now supports naive and ulp
  someone promised no bugs, here is one, lol
  ns on lenet_res
  Add HDF5 file importer for convert
  little changes to reverse to previous working lenet res
  Fix yolo2 and tiny_yolo2 models, fix classify
  add a lenet res model for testing and fun
  Detection test can now run
  Train can now start, but seems super slow
  Annoying bug in scatter, index out of range
  ...

+++ mayo/estimate.py
@@ -265,7 +265,7 @@ class ResourceEstimator(object):
             return 1
         valids = sum(np.sum(g.astype(np.int32)) for g in gates)
         totals = sum(g.size for g in gates)
-        return valids / totals
+        return float(valids / totals)
 
     def _gate_for_node(self, node):
         """
